Crate lambda

Modules

church

Constants for common functions from Church’s formalism.

Enums

Expression

An expression in the lambda calculus.

LambdaError

An error with a lambda expression.

Functions

evaluate

Evaluate an expression and return the normalized result.

evaluate_no_normalization

Evaluate an expression that is already normalized, without post-normalization

evaluate_normalized

Evaluate an expression that is already normalized.

normalize_variables

Normalize variables in an expression to avoid name shadowing.

Type Definitions

ParseResult

A result of parsing a lambda term.